pyproject.toml
[project]
name = "campaign-maker"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.108.0",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.5",
]

[project.scripts]
cmaker = "cmaker.main:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
